src/backend/requirements.txt
aiohttp==3.11.14
azure-cosmos==4.9.0
fastapi==0.115.8
fastapi-cli==0.0.7
gunicorn==23.0.0
python-dotenv==1.0.1
semantic-kernel==1.20.0
azure-search-documents==11.5.2
azure-identity==1.19.0
openai==1.61.1
azure-storage-blob==12.25.1